# Speed-up time spent in generating shaped str keys